Refactor action market commands and selectors
